+++ shotmanager/handlers/sm_overlay_tools_handlers.py
@@ -62,13 +62,17 @@ def install_handler_for_shot(self, context):
         # )
 
 
-def toggle_overlay_tools_display(self, context):
+def toggle_overlay_tools_display(context):
     # print("  toggle_overlay_tools_display:  self.UAS_shot_manager_display_overlay_tools: ", self.UAS_shot_manager_display_overlay_tools)
-    if self.UAS_shot_manager_display_overlay_tools:
-        a = bpy.ops.uas_shot_manager.sequence_timeline("INVOKE_DEFAULT")
+    prefs = context.preferences.addons["shotmanager"].preferences
+    if context.window_manager.UAS_shot_manager_display_overlay_tools:
+        if prefs.toggle_overlays_turnOn_sequenceTimeline:
+            a = bpy.ops.uas_shot_manager.sequence_timeline("INVOKE_DEFAULT")
         # print(f"a: {a}")
-        bpy.ops.uas_shot_manager.interactive_shots_stack("INVOKE_DEFAULT")
-        pass
+
+        if prefs.toggle_overlays_turnOn_interactiveShotsStack:
+            bpy.ops.uas_shot_manager.interactive_shots_stack("INVOKE_DEFAULT")
+
         # bpy.ops.uas_shot_manager.draw_cameras_ui("INVOKE_DEFAULT")
     else:
         pass