+++ shotmanager/operators/shots.py
@@ -69,9 +69,9 @@ class UAS_ShotManager_SetCurrentShot(Operator):
         # change shot
         if not event.shift and not event.ctrl:
             if event.alt:
-                props.setCurrentShotByIndex(self.index, changeTime=False, area=context.area)
+                props.setCurrentShotByIndex(self.index, changeTime=False, source_area=context.area)
             else:
-                props.setCurrentShotByIndex(self.index, area=context.area)
+                props.setCurrentShotByIndex(self.index, source_area=context.area)
             props.setSelectedShotByIndex(self.index)
             currentShotChanged = True
 
@@ -86,7 +86,7 @@ class UAS_ShotManager_SetCurrentShot(Operator):
 
         # already handled in first condition
         # elif event.alt and not event.shift and not event.ctrl:
-        #     props.setCurrentShotByIndex(self.index, changeTime=False, area=context.area)
+        #     props.setCurrentShotByIndex(self.index, changeTime=False, source_area=context.area)
         #     props.setSelectedShotByIndex(self.index)
 
         else:
@@ -1286,33 +1286,35 @@ class UAS_ShotManager_UniqueCameras(Operator):
 
     @staticmethod
     def unique_cam_name(cam_name):
-        return utils.unique_object_name()
+        return utils.unique_object_name(cam_name)
 
     def execute(self, context):
         scene = context.scene
-        takes = get_takes(context.scene)
+        props = context.scene.UAS_shot_manager_props
+        takes = props.getTakes()
         new_cam_from_shots = dict()
         objects = bpy.data.objects
         for take in takes:
-            existing_cam = set()
+            existing_cameras = set()
             for shot in take.shots:
                 if shot.camera is None:
                     continue
-                cam = shot.camera.name
+                camName = shot.camera.name
 
-                if cam in existing_cam and cam in objects:
+                if camName in existing_cameras and camName in objects:
                     if shot.name in new_cam_from_shots:
                         shot.camera = new_cam_from_shots[shot.name]
                     else:
-                        cam_obj = context.scene.objects[cam]
-                        new_cam = cam_obj.copy()
-                        new_cam.name = self.unique_cam_name(f"{cam}_{shot.name}")
-                        scene.collection.objects.link(new_cam)
+                        cam_obj = context.scene.objects[camName]
+                        new_cam = utils.duplicateObject(cam_obj)
+                        new_cam.name = self.unique_cam_name(f"{camName}_{shot.name}")
+                        new_cam.color = (uniform(0, 1), uniform(0, 1), uniform(0, 1), 1)
+                        # scene.collection.objects.link(new_cam)
                         new_cam_from_shots[shot.name] = new_cam.name
                         shot.camera = new_cam
 
                 new_cam_from_shots[shot.name] = shot.camera
-                existing_cam.add(cam)
+                existing_cameras.add(camName)
 
         return {"FINISHED"}